fix: player movement states
no movement during dead state + gravity during defending

+++ Assets/_ScriptableObjects/Enemies/EnemyObject.cs
@@ -5,8 +5,8 @@ public class EnemyObject : ScriptableObject
 {
     public int MaxHealth;
     public int Damage;
-    public int AttackInterval;
-    public int AttackDistance;
+    public float AttackInterval;
+    public float AttackDistance;
 
     public float Speed;
     public float SpottingRadius;
